DOC: Bring ``IntendedFor`` heuristic table from SDCFlows
The heuristic table will not be pertinent at the level of SDCFlows,
as the workflow will be staged for each fieldmap in the structure and
will not then be applied to correct the data - that will be the
responsibility of the host workflow (i.e., *fMRIPrep* for the case
at hand).

Since this table will be removed in the context of nipreps/sdcflows#123,
I'm bringing it here temporarily.